cli/templates/agent/requirements.txt
ray[serve]
fastapi
python-dotenv